+++ web/cypress.config.js
@@ -7,7 +7,7 @@ module.exports = defineConfig({
     viewportWidth: 1920,
     viewportHeight: 1080,
     setupNodeEvents(on, config) {
-      // implement node event listeners here
+
     },
   },
 });